requirements.txt
pillow
aiofiles
freezegun
pydantic
pytest
snakeviz
coverage
pytest-cov
black
mypy
sphynx
flake8
pylint
build
setuptools
twine
requests
numba
memory-profiler
